refactor(services): own web session and skill state in state modules
Relocates the 14 mutable bindings that fused the 99-method hub in
web-agent-client.ts into two state modules plus a seeds module. Neither
state module exports a binding: callers reach the state through
accessors, which read the live binding on every call and so cannot fork
into two copies of the mock world.

Method bodies are unchanged. The accessor set is cut to what the call
sites need, and every accessor has a caller outside its owning module.

web-agent-client.ts 3861 -> 3502.
